Bugfix 'resume_and_finish_from_inspection fn'. Reimplement it in 'syspect.Inspector' code.
This commit is probably too large.
EventAction.NORMAL was re-added as we now have a use for detecting an early return vs a normal return.
next_event has had the bulk of its code moved to a new function: 'handle_event'.
Examples were updated to work with and make use of new changes.