Allow user to inspect syscall result before resuming tracee
The main purpose of this is to allow easier detaching from tracees. To
detach, the tracee must be in a ptrace-stop state. We can now cleanly detach
after a syscall finishes.
This does add a new requirement for the user. They must call 'Inspector.resume_tracee' after a syscall finishes.